Refactor block testing data to own file. Fix lint.
